The tokenizer now returns one token at a time.
The parser will ask the tokenizer when it needs another. Then background for this change is that some times the parser need to change the state of the tokenizer before continuing, which is rather hard when all tokens have been created.

Also move all related classes to a parsing package.