Support &nbsp; character reference.
Before introducing support for ampersand (&), named character reference were tokenized as individual characters. Now, we instead read them as one and try to look up which character it represents and insert that single character instead.

So for the table consists of &nbsp;, will expand this based on what I run into in the wild.