Various extensions in parser for more spec compliance.
Triage a case where character references in attributes end up in ambigous ampersand state and start emitting character tokens. The parser does not expect this and ends up in a weird state. When fixing the ignored test, they should be added to the attribute value instead which should resolve the issue.